refactor(web): split feature modules
Move feature UI and optimistic update adapters under feature folders so shared optimistic helpers stay generic.
