+++ apps/web/convex/_generated/api.d.ts
@@ -13,11 +13,16 @@ import type * as auth from "../auth.js";
 import type * as dashboard from "../dashboard.js";
 import type * as http from "../http.js";
 import type * as items from "../items.js";
+import type * as lib_areaProjects from "../lib/areaProjects.js";
+import type * as lib_healthStatus from "../lib/healthStatus.js";
 import type * as lib_helpers from "../lib/helpers.js";
+import type * as lib_inboxProcessing from "../lib/inboxProcessing.js";
 import type * as lib_patch from "../lib/patch.js";
+import type * as lib_projectChanges from "../lib/projectChanges.js";
 import type * as lib_slugs from "../lib/slugs.js";
 import type * as lib_types from "../lib/types.js";
 import type * as lib_validation from "../lib/validation.js";
+import type * as lib_validators from "../lib/validators.js";
 import type * as projectLogs from "../projectLogs.js";
 import type * as projects from "../projects.js";
 
@@ -33,11 +38,16 @@ declare const fullApi: ApiFromModules<{
   dashboard: typeof dashboard;
   http: typeof http;
   items: typeof items;
+  "lib/areaProjects": typeof lib_areaProjects;
+  "lib/healthStatus": typeof lib_healthStatus;
   "lib/helpers": typeof lib_helpers;
+  "lib/inboxProcessing": typeof lib_inboxProcessing;
   "lib/patch": typeof lib_patch;
+  "lib/projectChanges": typeof lib_projectChanges;
   "lib/slugs": typeof lib_slugs;
   "lib/types": typeof lib_types;
   "lib/validation": typeof lib_validation;
+  "lib/validators": typeof lib_validators;
   projectLogs: typeof projectLogs;
   projects: typeof projects;
 }>;
