+++ apps/web/src/features/areas/components/area-card.tsx
