+++ apps/web/src/features/dashboard/components/attention-section.tsx
