+++ apps/web/src/features/projects/optimistic.ts
@@ -2,16 +2,10 @@ import { api } from "@convex/_generated/api";
 import type { Doc, Id } from "@convex/_generated/dataModel";
 import { nullsToUndefined } from "@convex/lib/patch";
 import type { OptimisticLocalStore } from "convex/browser";
+import { nextOrder, patchById, removeById } from "@/features/shared/optimistic";
 
-type Area = Doc<"areas">;
 type Project = Doc<"projects">;
 type ActionQueueItem = { id: string; text: string };
-
-type CreateAreaArgs = {
-  name: string;
-  standard?: string;
-  healthStatus: Area["healthStatus"];
-};
 
 type CreateProjectArgs = {
   name: string;
@@ -23,53 +17,12 @@ type NullablePatch<T> = {
   [K in keyof T]?: T[K] | null;
 };
 
-type AreaPatch = NullablePatch<
-  Pick<Area, "name" | "standard" | "healthStatus">
->;
-
 type ProjectPatch = NullablePatch<
   Pick<
     Project,
     "name" | "definitionOfDone" | "areaId" | "status" | "actionQueue" | "state"
   >
 >;
-
-export function patchById<T extends { _id: string }>(
-  items: T[],
-  id: string,
-  patch: Partial<T>,
-): T[] {
-  return items.map((item) => (item._id === id ? { ...item, ...patch } : item));
-}
-
-export function removeById<T extends { _id: string }>(
-  items: T[],
-  id: string,
-): T[] {
-  return items.filter((item) => item._id !== id);
-}
-
-export function completeNextAction<
-  T extends { actionQueue?: ActionQueueItem[] },
->(project: T): T {
-  return { ...project, actionQueue: (project.actionQueue ?? []).slice(1) };
-}
-
-export function buildOptimisticArea(
-  args: CreateAreaArgs,
-  options: { id: Id<"areas">; now: number; order: number },
-): Area {
-  return {
-    _id: options.id,
-    _creationTime: options.now,
-    userId: "",
-    name: args.name,
-    standard: args.standard,
-    healthStatus: args.healthStatus,
-    order: options.order,
-    createdAt: options.now,
-  };
-}
 
 export function buildOptimisticProject(
   args: CreateProjectArgs,
@@ -88,71 +41,10 @@ export function buildOptimisticProject(
   };
 }
 
-export function nextOrder(items: Array<{ order: number }>): number {
-  return items.reduce((max, item) => Math.max(max, item.order), -1) + 1;
-}
-
-export function optimisticallyCreateArea(
-  localStore: OptimisticLocalStore,
-  args: CreateAreaArgs,
-): void {
-  const current = localStore.getQuery(api.areas.list, {});
-  if (current === undefined) return;
-
-  localStore.setQuery(api.areas.list, {}, [
-    ...current,
-    buildOptimisticArea(args, {
-      id: crypto.randomUUID() as Id<"areas">,
-      now: Date.now(),
-      order: nextOrder(current),
-    }),
-  ]);
-}
-
-export function optimisticallyUpdateArea(
-  localStore: OptimisticLocalStore,
-  args: { id: Id<"areas"> } & AreaPatch,
-  options: { areaSlug: string },
-): void {
-  const { id, ...updates } = args;
-  const patch = nullsToUndefined(updates);
-
-  const current = localStore.getQuery(api.areas.list, {});
-  if (current !== undefined) {
-    localStore.setQuery(api.areas.list, {}, patchById(current, id, patch));
-  }
-
-  const single = localStore.getQuery(api.areas.get, { id });
-  if (single !== undefined && single !== null) {
-    localStore.setQuery(api.areas.get, { id }, { ...single, ...patch });
-  }
-
-  const bySlug = localStore.getQuery(api.areas.getBySlug, {
-    slug: options.areaSlug,
-  });
-  if (bySlug !== undefined && bySlug !== null) {
-    localStore.setQuery(
-      api.areas.getBySlug,
-      { slug: options.areaSlug },
-      {
-        ...bySlug,
-        ...patch,
-      },
-    );
-  }
-}
-
-export function optimisticallyRemoveArea(
-  localStore: OptimisticLocalStore,
-  args: { id: Id<"areas"> },
-  options: { areaSlug: string },
-): void {
-  const current = localStore.getQuery(api.areas.list, {});
-  if (current !== undefined) {
-    localStore.setQuery(api.areas.list, {}, removeById(current, args.id));
-  }
-  localStore.setQuery(api.areas.get, { id: args.id }, null);
-  localStore.setQuery(api.areas.getBySlug, { slug: options.areaSlug }, null);
+export function completeNextAction<
+  T extends { actionQueue?: ActionQueueItem[] },
+>(project: T): T {
+  return { ...project, actionQueue: (project.actionQueue ?? []).slice(1) };
 }
 
 export function optimisticallyCreateProjectInList(
@@ -237,9 +129,11 @@ export function optimisticallyRemoveProject(
       areaId: options.areaId,
     });
     if (areaProjects !== undefined) {
-      localStore.setQuery(api.projects.listByArea, { areaId: options.areaId }, [
-        ...removeById(areaProjects, args.id),
-      ]);
+      localStore.setQuery(
+        api.projects.listByArea,
+        { areaId: options.areaId },
+        removeById(areaProjects, args.id),
+      );
     }
   }
